renamed method constantKind to kind for dynamic_global_constants

+++ db/migration/1576502946543_createDynamicGlobalConstants.js
@@ -9,7 +9,7 @@ const dbKind = DbKindConstant.sqlDbKind;
 const upQuery =
   'CREATE TABLE `dynamic_global_constants` (\
   `id` bigint(20) NOT NULL AUTO_INCREMENT,\
-  `kind` int(11) NOT NULL,\
+  `kind` tinyint(4) NOT NULL,\
   `value` varchar(50) CHARACTER SET utf8 COLLATE utf8_unicode_ci NOT NULL,\
   `created_at` int(11) NOT NULL,\
   `updated_at` int(11) NOT NULL,\
